added routes folder
- organised routes into different routers
